docs added

+++ lib/features/appointement_scheduler/cubit/appointement_cubit.dart
@@ -11,12 +11,17 @@ class AppointmentCubit extends Cubit<AppointmentState> {
 
   final AppointmentService _appointmentService;
 
+  /// Adds an [Appointment].
   ///
+  /// This method sets the state to [AppointmentRemovingInProgress] and
+  /// checks if appointment is available.
+  /// [AppointmentBookingSuccess] is yielded after successful booking. In case
+  /// of an error or an unavailable [AppointmentBookingFailure] is yielded.
   Future<void> bookAppointment(Appointment appointment) async {
     emit(AppointmentBookingInProgress());
 
     try {
-      if(await _appointmentService.isAvailable(appointment)){
+      if (await _appointmentService.isAvailable(appointment)) {
         final code = await _appointmentService.bookAppointment(appointment);
         emit(AppointmentBookingSuccess(code));
       } else {
@@ -31,18 +36,27 @@ class AppointmentCubit extends Cubit<AppointmentState> {
     }
   }
 
+  /// Remove an appointment from database.
   ///
+  /// This method sets the state to [AppointmentRemovingInProgress] and to
+  /// [AppointmentRemovingSuccess] after successful removal and
+  /// [AppointmentRemovingFailure] if it is unsuccessful.
   Future<void> removeAppointment(String code) async {
     emit(AppointmentRemovingInProgress());
-
-    try {
-      await _appointmentService.removeAppointment(code);
-      emit(AppointmentRemovingSuccess());
-    } catch (error) {
-      emit(AppointmentRemovingFailure('Removing appointment failed.'));
+    if (code.isNotEmpty) {
+      try {
+        await _appointmentService.removeAppointment(code);
+        emit(AppointmentRemovingSuccess());
+      } catch (error) {
+        emit(AppointmentRemovingFailure('Removing appointment failed.'));
+      }
+    } else {
+      emit(AppointmentRemovingFailure('No code found'));
     }
   }
 
+  /// Reset the state to [AppointmentInitialLoading].
+  ///
   void resetState() {
     emit(AppointmentInitialLoading());
   }
